Warnings fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@12655 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_rpc/tapi_rpcsock_macros.h
@@ -940,7 +940,7 @@
     do {                                                                \
         int err_ = RPC_ERRNO(rpcs_);                                    \
                                                                         \
-        if (err_ != (exp_errno_))                                       \
+        if (err_ != (int)(exp_errno_))                                  \
         {                                                               \
             if (!!((#args_)[0]))                                        \
             {                                                           \
